Move 'classifierFMethodWithOptionallyExcludedPath' from Prometheus to MetricOps.
Addresses review comment: https://github.com/http4s/http4s/pull/3167#pullrequestreview-356424098.
